+++ ProyectoFinalJava/src/Models/Mecanico.java
@@ -10,7 +10,7 @@ public class Mecanico extends Persona implements IJson {
     private String contrasenia;
 
 
-    public Mecanico(String nombre, String apellido, int dni, int telefono, String email, String usuario, String contrasenia) {
+    public Mecanico(String nombre, String apellido, long dni, long telefono, String email, String usuario, String contrasenia) {
         super(nombre, apellido, dni, telefono, email);
         this.usuario = usuario;
         this.contrasenia = Seguridad.hashearContrasenia(contrasenia);
